Use paginated call to find users.
When there are more than 50 users, not all users show in the first page.
My user happened to be below the fold, which made the CLI unusable.
It checks if your user-id is in the list, but it would never be there.

This change forces the users call to do a paginated call
where it fetches all the pages.